move edge.hpp -> edge/edge.hpp
