+++ test/aoj/grl_1_a/static-graph.test.cpp
@@ -5,7 +5,7 @@
 #include "src/cpp-template/header/input.hpp"
 #include "src/cpp-template/header/rep.hpp"
 #include "src/cpp-template/header/type-alias.hpp"
-#include "src/graph/class/edge.hpp"
+#include "src/graph/class/edge/edge.hpp"
 #include "src/graph/single-source-shortest-path/in-non-negative-weighted-graph.hpp"
 
 #include <iostream>
