+++ test/aoj/grl_1_b/dynamic-graph.test.cpp
@@ -4,7 +4,7 @@
 
 #include "src/cpp-template/header/rep.hpp"
 #include "src/cpp-template/header/type-alias.hpp"
-#include "src/graph/class/edge.hpp"
+#include "src/graph/class/edge/edge.hpp"
 #include "src/graph/single-source-shortest-path/in-weighted-graph.hpp"
 
 #include <iostream>
