+++ test/aoj/grl_5_c/grl_5_c.test.cpp
@@ -2,7 +2,7 @@
 
 #include "src/cpp-template/header/rep.hpp"
 #include "src/cpp-template/header/type-alias.hpp"
-#include "src/graph/class/edge.hpp"
+#include "src/graph/class/edge/edge.hpp"
 #include "src/graph/class/static-graph.hpp"
 #include "src/graph/tree/offline-query/offline-query-lowest-common-ancestor.hpp"
