+++ test/atcoder/abc258_e/online-algorithm.test.cpp
@@ -3,7 +3,7 @@
 #include "src/cpp-template/header/rep.hpp"
 #include "src/cpp-template/header/type-alias.hpp"
 #include "src/cpp-template/header/vector-ios.hpp"
-#include "src/graph/class/edge.hpp"
+#include "src/graph/class/edge/edge.hpp"
 #include "src/graph/class/static-graph.hpp"
 #include "src/graph/functional-graph/online-query/online-query-jump-on-functional-graph.hpp"
