+++ test/library-checker/jump_on_tree.test.cpp
@@ -3,7 +3,7 @@
 #include "src/cpp-template/header/fast-ios.hpp"
 #include "src/cpp-template/header/rep.hpp"
 #include "src/cpp-template/header/type-alias.hpp"
-#include "src/graph/class/edge.hpp"
+#include "src/graph/class/edge/edge.hpp"
 #include "src/graph/class/static-graph.hpp"
 #include "src/graph/tree/offline-query/offline-query-jump-on-tree.hpp"
